Minor fix to the script
Minor fix to the script that rebuilds the database tables. Longitude and latitude values in the station_information table were in decimal format which didn't play nicely with jsonify. These were converted to floats and fixed in the script to do the same thing should the database need to be rebuilt.